remotecontrol: move the current implementation in its own folder
