+++ src/plugins/score-plugin-remotecontrol/RemoteControl/Websockets/Scenario/Event.cpp
