+++ src/plugins/score-plugin-remotecontrol/RemoteControl/Websockets/Scenario/Scenario.cpp
