+++ src/plugins/score-plugin-remotecontrol/RemoteControl/Websockets/Scenario/State.cpp
