+++ src/plugins/score-plugin-remotecontrol/RemoteControl/Websockets/Scenario/State.hpp
@@ -1,7 +1,7 @@
 #pragma once
 #include <Scenario/Document/State/StateModel.hpp>
 
-#include <RemoteControl/DocumentPlugin.hpp>
+#include <RemoteControl/Websockets/DocumentPlugin.hpp>
 
 namespace RemoteControl
 {
