+++ src/plugins/score-plugin-remotecontrol/RemoteControl/Websockets/Scenario/Sync.hpp
@@ -1,7 +1,7 @@
 #pragma once
 #include <Scenario/Document/TimeSync/TimeSyncModel.hpp>
 
-#include <RemoteControl/DocumentPlugin.hpp>
+#include <RemoteControl/Websockets/DocumentPlugin.hpp>
 
 namespace RemoteControl
 {
